port to newer wx

+++ src/rbprocess.h
@@ -36,7 +36,7 @@ int RBprocess(int olength, float* data, Marker* marker, Sample* sample, int n=1)
 
 	RubberBand::RubberBandStretcher ts(44100, 1, 0, 1);
 	// map
-	std::map<unsigned int, unsigned int> fmap;
+	std::map<long unsigned int, long unsigned int> fmap;
 	for (int i = 0; i < length; i += 1024) {
 		float old = marker->nnew2new(marker->new2old(i / (float)length));
 		fmap[int(old * olength) - ostart] = int(i);
